requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
cloudscraper>=1.2.71
selenium>=4.15.0
webdriver-manager>=4.0.1
undetected-chromedriver>=3.5.4
twilio>=8.0.0